fix: unit test err in config_test.go (#93)
type-casting to uint16 is needed.

+++ cmd/fledgectl/cmd/config_test.go
@@ -34,6 +34,6 @@ func TestLoadConfig(t *testing.T) {
 	config, err = loadConfig(configFilePath)
 	assert.Nil(t, err)
 	assert.Equal(t, "localhost", config.ApiServer.Host)
-	assert.Equal(t, 10100, config.ApiServer.Port)
+	assert.Equal(t, uint16(10100), config.ApiServer.Port)
 	assert.Equal(t, "john", config.User)
 }
